# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Digital Inspection App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      parameters:
        - in: query
          name: venueId
          schema: { type: integer }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Asset" }
    post:
      operationId: createAsset
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAssetInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
  /assets/{id}:
    patch:
      operationId: updateAsset
      tags: [assets]
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAssetInput" }
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    Asset:
      type: object
      properties:
        id: { type: integer }
        venueId: { type: integer }
        venueName: { type: string }
        venueCode: { type: string }
        type: { type: string }
        label: { type: string }
        serial:
          type: ["string", "null"]
        installedAt:
          type: ["string", "null"]
          format: date
        status: { type: string }
        notes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
      required: [id, venueId, venueName, venueCode, type, label, serial, installedAt, status, notes, createdAt]
    CreateAssetInput:
      type: object
      properties:
        venueId: { type: integer }
        type: { type: string }
        label: { type: string }
        serial:
          type: ["string", "null"]
        installedAt:
          type: ["string", "null"]
          format: date
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [venueId, type, label]
    UpdateAssetInput:
      type: object
      properties:
        type: { type: string }
        label: { type: string }
        serial:
          type: ["string", "null"]
        installedAt:
          type: ["string", "null"]
          format: date
        status: { type: string }
        notes:
          type: ["string", "null"]
...