# lib/api-zod/src/generated/types/updateAssetInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Inspection App API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAssetInput {
  type?: string;
  label?: string;
  /** @nullable */
  serial?: string | null;
  /** @nullable */
  installedAt?: Date | null;
  status?: string;
  /** @nullable */
  notes?: string | null;
}